pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Phase 1 private Telegram bot with FastAPI, SQLite memory, and OpenAI replies."
readme = "docs/roadmap.md"
requires-python = ">=3.10"
dependencies = [
  "aiogram>=3.26.0,<4.0.0",
  "aiosqlite>=0.21.0,<1.0.0",
  "fastapi>=0.128.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "openai>=2.11.0,<3.0.0",
  "pydantic-settings>=2.11.0,<3.0.0",
  "uvicorn>=0.38.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.2.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
